Bug 575787 - [win32] Optimize Tree.setItemCount
On Windows, Tree maintains its items with a single linked list. For this
reason, "find previous" operation is O(N) because it requires walking
from start of the list.

When inserting an item, it needs to be linked into the list, which
involves finding the previous item and updating its 'next' field, which
is O(N) for one item and O(N*N) for inserting N items.

The workaround is to bulk insert all items at the same position. This
means that if TreeItem had no children, inserting N items into it is
O(N) instead of O(N*N).

On top of that, this patch preallocates 'Tree.items[]' to hold exactly
the required number of items.

Resulting performance improvement on my machine:
                |      before patch     |      after patch      |
                +-----------+-----------+-----------+-----------+
                |   REGULAR |   VIRTUAL |   REGULAR |   VIRTUAL |
----------------+-----------+-----------+-----------+-----------+
   10_000 items |   0,44sec |   0,43sec |   0,07sec |   0,06sec |
  100_000 items | 113,83sec | 111,95sec |   0,70sec |   0,63sec |
1_000_000 items |  too long |  too long |   6,82sec |   6,28sec |

Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.swt/+/187203
Tested-by: Niraj Modi <niraj.modi@in.ibm.com>

+++ bundles/org.eclipse.swt/Eclipse SWT/win32/org/eclipse/swt/widgets/Tree.java
@@ -2046,7 +2046,10 @@ void createItem (TreeColumn column, int index) {
 	}
 }
 
-// For fast bulk insert, see comments for TreeItem#TreeItem(TreeItem,int,int)
+/**
+ * The fastest way to insert many items is documented in {@link TreeItem#TreeItem(Tree,int,int)}
+ * and {@link TreeItem#setItemCount}
+ */
 void createItem (TreeItem item, long hParent, long hInsertAfter, long hItem) {
 	int id = -1;
 	if (item != null) {
@@ -2066,9 +2069,8 @@ void createItem (TreeItem item, long hParent, long hInsertAfter, long hItem) {
 				shrink = true;
 				length = Math.max (4, items.length * 3 / 2);
 			}
-			TreeItem [] newItems = new TreeItem [length];
-			System.arraycopy (items, 0, newItems, 0, items.length);
-			items = newItems;
+
+			itemsGrowArray (length);
 		}
 		lastID = id + 1;
 	}
@@ -3851,6 +3853,22 @@ boolean isUseWsBorder () {
 	return true;
 }
 
+int itemsGetFreeCapacity() {
+	int count = 0;
+	for (TreeItem item : items) {
+		if (item == null)
+			count++;
+	}
+
+	return count;
+}
+
+void itemsGrowArray (int newCapacity) {
+	TreeItem [] newItems = new TreeItem [newCapacity];
+	System.arraycopy (items, 0, newItems, 0, items.length);
+	items = newItems;
+}
+
 void redrawSelection () {
 	if ((style & SWT.SINGLE) != 0) {
 		long hItem = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_CARET, 0);
@@ -4113,6 +4131,9 @@ public void setInsertMark (TreeItem item, boolean before) {
 
 /**
  * Sets the number of root-level items contained in the receiver.
+ * <p>
+ * The fastest way to insert many items is documented in {@link TreeItem#TreeItem(Tree,int,int)}
+ * and {@link TreeItem#setItemCount}
  *
  * @param count the number of items
  *
@@ -4130,18 +4151,41 @@ public void setItemCount (int count) {
 }
 
 void setItemCount (int count, long hParent) {
-	long hItem = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_CHILD, hParent);
-
 	boolean redraw = false;
 	if (OS.SendMessage (handle, OS.TVM_GETCOUNT, 0, 0) == 0) {
 		redraw = getDrawing () && OS.IsWindowVisible (handle);
 		if (redraw) OS.DefWindowProc (handle, OS.WM_SETREDRAW, 0, 0);
 	}
-	int itemCount = 0;
-	while (hItem != 0 && itemCount < count) {
-		hItem = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_NEXT, hItem);
-		itemCount++;
+
+	// Investigate existing items and decide what to do
+	long itemInsertAfter = 0;
+	int  numInserted = 0;
+	long itemDeleteFrom = 0;
+	{
+		// Iterate to position #count and find prev/next items at this position
+		int itemCount = 0;
+		long itemPrev = OS.TVI_FIRST;
+		long itemNext = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_CHILD, hParent);
+		while (itemNext != 0 && itemCount < count)
+		{
+			itemPrev = itemNext;
+			itemNext = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_NEXT, itemNext);
+			itemCount++;
+		}
+
+		if ((itemCount == count) && (itemNext == 0)) {
+			// Exactly 'count' items, no need to do anything.
+			return;
+		} else if (itemCount == count) {
+			// Too many items, going to delete some
+			itemDeleteFrom = itemNext;
+		} else if (itemNext == 0) {
+			// Counted all items, and there is not enough, going to insert some.
+			itemInsertAfter = itemPrev;
+			numInserted = count - itemCount;
+		}
 	}
+
 	boolean expanded = false;
 	TVITEM tvItem = new TVITEM ();
 	tvItem.mask = OS.TVIF_HANDLE | OS.TVIF_PARAM;
@@ -4155,32 +4199,40 @@ void setItemCount (int count, long hParent) {
 		int state = (int)OS.SendMessage (handle, OS.TVM_GETITEMSTATE, hParent, OS.TVIS_EXPANDED);
 		expanded = (state & OS.TVIS_EXPANDED) != 0;
 	}
-	while (hItem != 0) {
-		tvItem.hItem = hItem;
-		OS.SendMessage (handle, OS.TVM_GETITEM, 0, tvItem);
-		hItem = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_NEXT, hItem);
-		TreeItem item = tvItem.lParam != -1 ? items [(int)tvItem.lParam] : null;
-		if (item != null && !item.isDisposed ()) {
-			item.dispose ();
-		} else {
-			releaseItem (tvItem.hItem, tvItem, false);
-			destroyItem (null, tvItem.hItem);
-		}
-	}
-	if ((style & SWT.VIRTUAL) != 0) {
-		for (int i=itemCount; i<count; i++) {
-			if (expanded) ignoreShrink = true;
-			createItem (null, hParent, OS.TVI_LAST, 0);
-			if (expanded) ignoreShrink = false;
+
+	if (itemDeleteFrom != 0) {
+		while (itemDeleteFrom != 0) {
+			tvItem.hItem = itemDeleteFrom;
+			OS.SendMessage (handle, OS.TVM_GETITEM, 0, tvItem);
+			itemDeleteFrom = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_NEXT, itemDeleteFrom);
+			TreeItem item = tvItem.lParam != -1 ? items [(int)tvItem.lParam] : null;
+			if (item != null && !item.isDisposed ()) {
+				item.dispose ();
+			} else {
+				releaseItem (tvItem.hItem, tvItem, false);
+				destroyItem (null, tvItem.hItem);
+			}
 		}
 	} else {
-		shrink = true;
-		int extra = Math.max (4, (count + 3) / 4 * 4);
-		TreeItem [] newItems = new TreeItem [items.length + extra];
-		System.arraycopy (items, 0, newItems, 0, items.length);
-		items = newItems;
-		for (int i=itemCount; i<count; i++) {
-			new TreeItem (this, SWT.NONE, hParent, OS.TVI_LAST, 0);
+		// For performance reasons, reserve the necessary space in items[]
+		int freeCapacity = itemsGetFreeCapacity();
+		if (numInserted > freeCapacity)
+			itemsGrowArray (items.length + numInserted - freeCapacity);
+
+		// Note: on Windows, insert complexity is O(pos), so for performance
+		// reasons, all items are inserted at minimum possible position, that
+		// is, all at the same position.
+
+		if ((style & SWT.VIRTUAL) != 0) {
+			for (int i = 0; i < numInserted; i++) {
+				if (expanded) ignoreShrink = true;
+				createItem (null, hParent, itemInsertAfter, 0);
+				if (expanded) ignoreShrink = false;
+			}
+		} else {
+			for (int i = 0; i < numInserted; i++) {
+				new TreeItem (this, SWT.NONE, hParent, itemInsertAfter, 0);
+			}
 		}
 	}
 	if (redraw) {
